+++ lib/appsignal/diagnose/agent.ex
@@ -5,10 +5,13 @@ defmodule Appsignal.Diagnose.Agent do
     if @nif.loaded? do
       Appsignal.Nif.env_put("_APPSIGNAL_DIAGNOSE", "true")
       report_string = @nif.diagnose
-      report = case Poison.decode(report_string) do
-        {:ok, report} -> {:ok, report}
-        {:error, _} -> {:error, report_string}
-      end
+
+      report =
+        case Poison.decode(report_string) do
+          {:ok, report} -> {:ok, report}
+          {:error, _} -> {:error, report_string}
+        end
+
       Appsignal.Nif.env_delete("_APPSIGNAL_DIAGNOSE")
       report
     else
@@ -20,33 +23,36 @@ defmodule Appsignal.Diagnose.Agent do
   # does go through the whole process of setting the config to the
   # environment.
   def print(report) do
-    IO.puts "Agent diagnostics"
+    IO.puts("Agent diagnostics")
+
     if report["error"] do
-      IO.puts "  Error: #{report["error"]}"
+      IO.puts("  Error: #{report["error"]}")
     else
-      Enum.each(report_definition(), fn({component, definition}) ->
-        IO.puts "  #{definition[:label]}"
+      Enum.each(report_definition(), fn {component, definition} ->
+        IO.puts("  #{definition[:label]}")
         print_component(report[component] || %{}, definition[:tests])
       end)
     end
-    IO.puts ""
+
+    IO.puts("")
   end
 
   defp print_component(report, categories) do
-    Enum.each(categories, fn({category, tests}) ->
+    Enum.each(categories, fn {category, tests} ->
       print_category(report[category] || %{}, tests)
     end)
   end
 
   defp print_category(report, tests) do
-    Enum.each(tests, fn({test, definition}) ->
+    Enum.each(tests, fn {test, definition} ->
       print_test(report[test] || %{}, definition)
     end)
   end
 
   defp print_test(report, definition) do
-    IO.write "    #{definition[:label]}: "
+    IO.write("    #{definition[:label]}: ")
     result = report["result"]
+
     display_value =
       case Map.has_key?(definition, :values) do
         true ->
@@ -54,16 +60,20 @@ defmodule Appsignal.Diagnose.Agent do
             {:ok, value} -> value
             :error -> nil
           end
-        false -> result
+
+        false ->
+          result
       end
+
     display_value =
       case display_value do
         value when value in [nil, ""] -> "-"
         value -> value
       end
-    IO.puts display_value
-    if report["error"], do: IO.puts "      Error: #{report["error"]}"
-    if report["output"], do: IO.puts "      Output: #{report["output"]}"
+
+    IO.puts(display_value)
+    if report["error"], do: IO.puts("      Error: #{report["error"]}")
+    if report["output"], do: IO.puts("      Output: #{report["output"]}")
   end
 
   defp report_definition do
@@ -74,7 +84,7 @@ defmodule Appsignal.Diagnose.Agent do
           "config" => %{
             "valid" => %{
               :label => "Configuration",
-              :values => %{ true => "valid", false => "invalid" }
+              :values => %{true => "valid", false => "invalid"}
             }
           }
         }
@@ -85,34 +95,34 @@ defmodule Appsignal.Diagnose.Agent do
           "boot" => %{
             "started" => %{
               :label => "Started",
-              :values => %{ true => "started", false => "not started" }
+              :values => %{true => "started", false => "not started"}
             }
           },
           "host" => %{
-            "uid" => %{ :label => "Process user id" },
-            "gid" => %{ :label => "Process user group id" }
+            "uid" => %{:label => "Process user id"},
+            "gid" => %{:label => "Process user group id"}
           },
           "config" => %{
             "valid" => %{
               :label => "Configuration",
-              :values => %{ true => "valid", false => "invalid" }
+              :values => %{true => "valid", false => "invalid"}
             }
           },
           "logger" => %{
             "started" => %{
               :label => "Logger",
-              :values => %{ true => "started", false => "not started" }
+              :values => %{true => "started", false => "not started"}
             }
           },
           "working_directory_stat" => %{
-            "uid" => %{ :label => "Working directory user id" },
-            "gid" => %{ :label => "Working directory user group id" },
-            "mode" => %{ :label => "Working directory permissions" }
+            "uid" => %{:label => "Working directory user id"},
+            "gid" => %{:label => "Working directory user group id"},
+            "mode" => %{:label => "Working directory permissions"}
           },
           "lock_path" => %{
             "created" => %{
               :label => "Lock path",
-              :values => %{ true => "writable", false => "not writable" }
+              :values => %{true => "writable", false => "not writable"}
             }
           }
         }