+++ lib/appsignal/system.ex
@@ -22,7 +22,9 @@ defmodule Appsignal.System do
           {int, _} -> int
           :error -> nil
         end
-      {_, _} -> nil
+
+      {_, _} ->
+        nil
     end
   end