+++ lib/appsignal/utils/tasks.ex
@@ -1,11 +1,11 @@
 defmodule :appsignal_tasks do
   def diagnose do
     Mix.Tasks.Appsignal.Diagnose.run(nil)
-    :init.stop
+    :init.stop()
   end
 
   def demo do
     Mix.Tasks.Appsignal.Demo.run(nil)
-    :init.stop
+    :init.stop()
   end
 end