+++ test/phoenix/phoenix_test.exs
@@ -3,7 +3,7 @@ defmodule UsingAppsignalPhoenix do
     conn |> Plug.Conn.assign(:called?, true)
   end
 
-  defoverridable [call: 2]
+  defoverridable call: 2
 
   use Appsignal.Phoenix
 end
@@ -13,7 +13,7 @@ defmodule UsingAppsignalPhoenixWithException do
     raise "exception!"
   end
 
-  defoverridable [call: 2]
+  defoverridable call: 2
 
   use Appsignal.Phoenix
 end
@@ -23,7 +23,7 @@ defmodule UsingAppsignalPhoenixWithTimeout do
     Task.async(fn -> :timer.sleep(10) end) |> Task.await(1)
   end
 
-  defoverridable [call: 2]
+  defoverridable call: 2
 
   use Appsignal.Phoenix
 end
@@ -35,7 +35,7 @@ defmodule Appsignal.PhoenixTest do
   alias Appsignal.{Transaction, FakeTransaction}
 
   setup do
-    {:ok, fake_transaction} = FakeTransaction.start_link
+    {:ok, fake_transaction} = FakeTransaction.start_link()
     [fake_transaction: fake_transaction]
   end
 
@@ -61,47 +61,58 @@ defmodule Appsignal.PhoenixTest do
 
   describe "concerning catching errors" do
     setup do
-      conn = %Plug.Conn{}
-      |> Plug.Conn.put_private(:phoenix_controller, "foo")
-      |> Plug.Conn.put_private(:phoenix_action, "bar")
+      conn =
+        %Plug.Conn{}
+        |> Plug.Conn.put_private(:phoenix_controller, "foo")
+        |> Plug.Conn.put_private(:phoenix_action, "bar")
 
       [conn: conn]
     end
 
     test "reports an error for an exception", %{conn: conn, fake_transaction: fake_transaction} do
-      :ok = try do
-        UsingAppsignalPhoenixWithException.call(conn, %{})
-      catch
-        :error, %RuntimeError{message: "exception!"} -> :ok
-        type, reason -> {type, reason}
-      end
+      :ok =
+        try do
+          UsingAppsignalPhoenixWithException.call(conn, %{})
+        catch
+          :error, %RuntimeError{message: "exception!"} -> :ok
+          type, reason -> {type, reason}
+        end
 
       assert FakeTransaction.started_transaction?(fake_transaction)
-      assert [{
-        %Appsignal.Transaction{} = transaction,
-        "RuntimeError",
-        "HTTP request error: exception!",
-        _stack
-      }] = FakeTransaction.errors(fake_transaction)
+
+      assert [
+               {
+                 %Appsignal.Transaction{} = transaction,
+                 "RuntimeError",
+                 "HTTP request error: exception!",
+                 _stack
+               }
+             ] = FakeTransaction.errors(fake_transaction)
+
       assert [transaction] == FakeTransaction.finished_transactions(fake_transaction)
       assert [transaction] == FakeTransaction.completed_transactions(fake_transaction)
     end
 
     test "reports an error for a timeout", %{conn: conn, fake_transaction: fake_transaction} do
-      :ok = try do
-        UsingAppsignalPhoenixWithTimeout.call(conn, %{})
-      catch
-        :exit, {:timeout, {Task, :await, _}} -> :ok
-        type, reason -> {type, reason}
-      end
+      :ok =
+        try do
+          UsingAppsignalPhoenixWithTimeout.call(conn, %{})
+        catch
+          :exit, {:timeout, {Task, :await, _}} -> :ok
+          type, reason -> {type, reason}
+        end
 
       assert FakeTransaction.started_transaction?(fake_transaction)
-      assert [{
-        %Appsignal.Transaction{} = transaction,
-        ":timeout",
-        "HTTP request error: {:timeout, {Task, :await, [%Task{owner: " <> _,
-        _stack
-      }] = FakeTransaction.errors(fake_transaction)
+
+      assert [
+               {
+                 %Appsignal.Transaction{} = transaction,
+                 ":timeout",
+                 "HTTP request error: {:timeout, {Task, :await, [%Task{owner: " <> _,
+                 _stack
+               }
+             ] = FakeTransaction.errors(fake_transaction)
+
       assert [transaction] == FakeTransaction.finished_transactions(fake_transaction)
       assert [transaction] == FakeTransaction.completed_transactions(fake_transaction)
     end